Use [u]intN_t for Nintendo 64 data

+++ src/r4300/cp1.h
@@ -22,14 +22,16 @@
 #ifndef M64P_R4300_CP1_H
 #define M64P_R4300_CP1_H
 
+#include <stdint.h>
+
 extern float *reg_cop1_simple[32];
 extern double *reg_cop1_double[32];
-extern int FCR0, FCR31;
-extern long long int reg_cop1_fgr_64[32];
+extern int32_t FCR0, FCR31;
+extern int64_t reg_cop1_fgr_64[32];
 extern int rounding_mode, trunc_mode, round_mode, ceil_mode, floor_mode;
 
-void shuffle_fpr_data(int oldStatus, int newStatus);
-void set_fpr_pointers(int newStatus);
+void shuffle_fpr_data(uint32_t oldStatus, uint32_t newStatus);
+void set_fpr_pointers(uint32_t newStatus);
 
 #endif /* M64P_R4300_CP1_H */
